jtag: add stlink in parallel to jlink
now I have two bords, one with jlink (RMP) and
one with st-link 8stm32dicovery). To be able to flash both
it is necessary to ad a second make target to flash both.
make flash -> RMP Panel with jlink
make stflash -> STM32DISCOVERY panel with st-link

The st-link may have V2 or V2.1, it must be changed in
/usr/share/openocd/scripts/board/stm32f4discovery.cfg